requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.26.4
azure-storage-blob==12.16.0
boto3==1.26.100
botocore==1.29.100
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.1.7
django-better-admin-arrayfield==1.4.2
django-storages==1.13.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
jmespath==1.0.1
msgpack==1.0.5
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzoom==1.0.3
redis==4.6.0
requests==2.30.0
s3transfer==0.6.0
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.3.0
zope.interface==6.0
